LayoutTests/http/tests/media/hls/video-controller-getStartDate-expected.txt

 
Test that getStartDate() returns appropriate NaN or date.

EVENT(canplaythrough)
EVENT(canplaythrough)
EXPECTED (isNaN(video.getStartDate()) == 'true') OK
EXPECTED (video.getStartDate() == 'Wed Nov 03 2010 01:00:00 GMT-0700 (PDT)') OK
END OF TEST
